backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
openai==1.6.1
anthropic==0.7.1
aiohttp==3.9.1
supabase==2.0.3
postgrest-py==0.13.2

